pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]


[project]
name = "Job_Match"
version = "0.0.1"
description = "An application that can match companies with professionals that are looking for a job."
classifiers = ["Private :: Do Not Upload", "Programming Language :: Python :: 3.12"]
readme = "README.md"
requires-python = ">=3.12"


dependencies = [
    "fastapi==0.114.2",
    "uvicorn[standard]==0.32.0",
    "SQLAlchemy==2.0.34",
    "python-jose==3.3.0",
    "psycopg2==2.9.9",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "email-validator==2.2.0",
    "python-multipart==0.0.12",
    "pydantic-settings==2.6.1",
    "ecs-logging==2.2.0",
    "types-passlib==1.7.7.20240819",
    "types-python-jose==3.3.4.20240106",
    "debugpy==1.8.9",
    "itsdangerous==2.2.0",
    "requests==2.32.3",
    "types-requests==2.32.0.20241016",
    "httpx==0.27.2",
    "pytest-asyncio==0.24.0"
]

[project.optional-dependencies]
test = [
    "autoflake==2.2.1", 
    "black==23.9.1", 
    "isort[colors]==5.8.0", 
    "mypy==1.5.1", 
    "pylint==3.0.0a7",
    "pytest==8.2.0", 
    "pytest-mock==3.14.0", 
    "pytest-cov==4.1.0", 
    "coverage==7.3.1"
]

[project.urls]
Source = "https://github.com/Forum-System-Developers/job-match"

[tool.black]
exclude = '(\.git|\.mypy_cache|\.venv)'
include = '\.pyi?$'
target-version = ['py311']
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "--cache-clear --log-cli-level=DEBUG --strict-markers --cov-branch --cov-fail-under=30"
junit_family = "xunit2"
log_format = "%(asctime)s %(levelname)s %(name)s %(message)s"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration with other components",
    "api: marks tests as api tests",
    "asyncio: mark a test as asyncio-based."
]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.coverage.report]
show_missing = true
fail_under = 50
omit = [
  "**/tests/*",
  "**/src/mm_reprocess/*",
  "**/src/config.py",
  "**/src/profiler_utils.py",
]

[tool.coverage.run]
source = ["src"]

[tool.mypy]
plugins = "pydantic.mypy"
